+++ src/ApiProxy/ApiProxyInterface.php
@@ -5,16 +5,20 @@ namespace CaliforniaMountainSnake\LongmanTelegrambotLaravelApiAuthSystem\ApiProx
 use CaliforniaMountainSnake\LongmanTelegrambotLaravelApiAuthSystem\ApiProxy\Exceptions\ApiProxyException;
 use CaliforniaMountainSnake\UtilTraits\Curl\HttpResponse;
 
+/**
+ * The interface that allows object to perform requests to API routes.
+ */
 interface ApiProxyInterface
 {
     /**
-     * Execute query to the target api route.
+     * Execute a query to the target api route.
      *
-     * @param AvailableRoute $_rote   Route.
-     * @param array          $_params Query parameters [optional].
+     * @param AvailableRoute $_rote    Route.
+     * @param array          $_params  Query parameters [optional].
+     * @param array|null     $_headers Query headers [optional].
      *
      * @return HttpResponse
      * @throws ApiProxyException
      */
-    public function query(AvailableRoute $_rote, array $_params = []): HttpResponse;
+    public function query(AvailableRoute $_rote, array $_params = [], array $_headers = null): HttpResponse;
 }